fix: reorganized code and commented simulation_main
