+++ python-server/server-master/Python/server_utils/trial_result.py
@@ -4,8 +4,8 @@
 # is possible to obtain a response vector made of 1 (in case of correct answers)
 # and 0 (in case of incorrect answers)
 
-from area_data import AreaData
-from trial import Trial
+from server_utils.area_data import AreaData
+from server_utils.trial import Trial
 class TrialResult:
 
     def __init__(self, difficulty, decision_time, correct, mode, raw_trial_data = None, trial_data = None):
