+++ python-server/server-master/Python/test/__init__.py
@@ -1,0 +1,4 @@
+#RUN THE TESTS FROM PYTHON DIRECTORY: python -m test.tests
+
+#import thesis
+#import tests
